Migrate build pipeline to GitHub Actions
This commit mogrates the build pipeline from Codebuild to GitHub
actions. The change is equivalent except for adding 2 new resources that
were added recently (DBParameterGroup and DBOptionGroup). The former is
disabled as of yet due to an insufficient code coverage score (TODO).

The pipeline uses OpenJDK8 and Python 3.7.

Newly implemented resources will not be added to the build pipeline
automatically (at least yet). Instead, they are implemented as
independent pipeline steps. This approach immediately narrows down a
pipeline failure to a specific resource.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>