AWS::RDS::OptionGroup supports system tags
This commit adds a support for system tags. Prior to that OptionGroup
was supporting resource-level tags only. RDS API makes no distinction
between system- and resource-level tags, so the tags are merged into a
single collection before invoking the corresponding RDS API call.

Apart from this change, OptionGroup drops a support for tagging
soft-failing. As it was concluded internally, soft failing on tagging is
considered to be a deprecated mode, therefore it is not making its way
to the new implementation.

AWS SDK version was bumped to 2.16.54.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>